requirements.txt
django==2.2.7
djangorestframework==3.12.4
gunicorn==19.0.0
pillow==8.1.0
django-cors-headers==3.8.0
psycopg2==2.8.6